feat, refactor: implement basic field motion, parameters, and GUI feedback - Add field parameters to backend + controls to editor - Implement updateField() with RING, RANDOM_DISCRETE, and RANDOM_WALK patterns 	- Extract randomDiscrete() and randomWalk() patterns into static methods 	- Combine MotionController parameters and state into structs, clean up setters / getters 	- Separate timekeeping into positionTime and fieldTime, advance phase independently - Implement live visualization of field coordinates in PolarMapComponent - Extract toBounds() helper for drawing indicator circles from coordinates - Change rate controls to bipolar

+++ Source/ConvolutionReverb.cpp
@@ -193,7 +193,7 @@ void ConvolutionReverb::setDecay(float decay) {
 	float alpha = totalDB / static_cast<float>(maxIRPartitionCount); // dB
 	float envelope = powf(10.0f, alpha / 20.0f); // Linear
 	for (int partition = 0; partition < maxIRPartitionCount; ++partition)
-		irEnvelopes[partition] = powf(envelope, partition);
+		irEnvelopes[partition] = powf(envelope, static_cast<float>(partition));
 }
 
 void ConvolutionReverb::process(juce::AudioBuffer<float>& in) {
